source/requirements.txt
openai
langchain
langchain_community
langchain_openai
pinecone-client
python-dotenv
tiktoken
pypdf
langchain-pinecone